# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MissionOut API",
    "description": "MissionOut API contract for dispatcher, team admin, and responder clients. The exported OpenAPI document in contracts/openapi.json is the cross-stack source of truth.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/incidents": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "List Incidents",
        "operationId": "list_incidents_incidents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentRead"
                  },
                  "type": "array",
                  "title": "Response List Incidents Incidents Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Create Incident",
        "operationId": "create_incident_incidents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/incidents/{incident_public_id}": {
      "patch": {
        "tags": [
          "incidents"
        ],
        "summary": "Update Incident",
        "operationId": "update_incident_incidents__incident_public_id__patch",
        "parameters": [
          {
            "name": "incident_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Public Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/incidents/{incident_public_id}/responses": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Create Incident Response",
        "operationId": "create_incident_response_incidents__incident_public_id__responses_post",
        "parameters": [
          {
            "name": "incident_public_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Incident Public Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponseRecordCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseRecordRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "team_public_id": {
            "type": "string",
            "title": "Team Public Id"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "title",
          "team_public_id",
          "location",
          "notes"
        ],
        "title": "IncidentCreate"
      },
      "IncidentRead": {
        "properties": {
          "public_id": {
            "type": "string",
            "title": "Public Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "team_public_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Public Id"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "title": "Created"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "responses": {
            "items": {
              "$ref": "#/components/schemas/ResponseRecordRead"
            },
            "type": "array",
            "title": "Responses"
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "title",
          "location",
          "created",
          "notes",
          "active",
          "responses"
        ],
        "title": "IncidentRead"
      },
      "IncidentUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "title",
          "location",
          "notes",
          "active"
        ],
        "title": "IncidentUpdate"
      },
      "ResponseRecordCreate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "rank": {
            "type": "integer",
            "title": "Rank",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "status",
          "source"
        ],
        "title": "ResponseRecordCreate"
      },
      "ResponseRecordRead": {
        "properties": {
          "user_public_id": {
            "type": "string",
            "title": "User Public Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "title": "Updated"
          }
        },
        "type": "object",
        "required": [
          "user_public_id",
          "status",
          "rank",
          "updated"
        ],
        "title": "ResponseRecordRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}